refactor: RemoteEvaluateDataSource.kt 인터페이스 분리 원칙(isp) 적용
