+++ data/src/main/java/com/suwiki/data/datasource/remote/RemoteEvaluateReportDataSource.kt
@@ -1,0 +1,11 @@
+package com.suwiki.data.datasource.remote
+
+import com.suwiki.model.Result
+
+interface RemoteEvaluateReportDataSource {
+
+    suspend fun reportLecture(
+        evaluateIdx: Long,
+        content: String = "",
+    ): Result<Unit>
+}
